Update prayerTimesBot.py
Notifications not starting automatically after midnight:

To fix this, we can add a job that runs every day at midnight to trigger the daily update automatically. This way, you won't need to send /start to get notifications.

Duplicate notifications:

This is likely happening because the notifications are being scheduled multiple times. We need to clear the existing jobs before scheduling new ones.